# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Holiday Planner API",
        "description": "This is the API documentation for the Holiday Planner application.",
        "contact": {
            "email": "lucasabrandaoo@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/holiday-plan/{id}/pdf": {
            "get": {
                "tags": [
                    "Holiday Plans"
                ],
                "summary": "Generate Holiday Plan PDF",
                "description": "Generate and download a PDF file of the holiday plan with the specified ID.",
                "operationId": "generateHolidayPlanPDF",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the holiday plan to generate the PDF for",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF file generated successfully",
                        "content": {
                            "application/pdf": {
                                "schema": {
                                    "description": "PDF binary stream",
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Holiday Plan not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message when the plan is not found",
                                            "type": "string",
                                            "example": "Holiday Plan not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message indicating the user is not authenticated",
                                            "type": "string",
                                            "example": "Unauthorized."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Generic server error message",
                                            "type": "string",
                                            "example": "Internal server error."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
...
}